+++ CHANGES.rst
@@ -1,7 +1,9 @@
 
-Revision 0.1.5, XX-10-2017
+Revision 0.1.5, XX-11-2017
 --------------------------
 
+- Added MIB *status*, *release* and *revision description* set calls
+  at pysnmp code generator
 - Sphinx documentation theme changed to Alabaster
 
 Revision 0.1.4, 14-10-2017